Перенесённые функции для tab1
